+++ CHANGELOG.md
@@ -5,11 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 0.2.0 - [Unreleased]
+## 0.2.0 [Unreleased]
 
 ### Added
 
-* [#8] **Inspection**: *column without annotation* now try to guess the column type (read the [doc](https://github.com/rentalhost/laravel-insight/wiki/Inspections#column-without-annotation) to more information);
+* [#10] **Query Scopes**: *ctrl+click* on a scoped method resolve to scope declaration ([more info](https://github.com/rentalhost/laravel-insight/wiki/Auxiliary-Features#query-scopes));
+* [#10] **Query Scopes**: code completion to scope methods ([more info](https://github.com/rentalhost/laravel-insight/wiki/Auxiliary-Features#query-scopes));
+* [#8] **Inspection**: *column without annotation* now try to guess the column type ([more info](https://github.com/rentalhost/laravel-insight/wiki/Inspections#column-without-annotation));
 
 ### Fixed
 
@@ -30,5 +32,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 * Initial development release;
-* **Inspection:** column without annotation ([doc](https://github.com/rentalhost/laravel-insight/wiki/Inspections#column-without-annotation));
+* **Inspection:** column without annotation ([more info](https://github.com/rentalhost/laravel-insight/wiki/Inspections#column-without-annotation));